requirements.txt
asgiref==3.3.1
cryptography==3.3.1
Django==3.1.5
django-allauth==0.44.0
django-appconf==1.0.4
django-crispy-forms==1.10.0
django-embed-video==1.4.0
oauthlib==3.1.0
Pillow==8.1.0
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
rcssmin==1.0.6
requests-oauthlib==1.3.0
rjsmin==1.1.0
sqlparse==0.4.1
